Breaking up server Express routes functions.
